pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "ml-training-operator"
version = "0.1.0"
description = "Kubernetes-native ML training job operator for autonomous vehicle perception models"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Abhinav Yedla", email = "lsayedla@gmail.com" },
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "kubernetes>=28.1.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "httpx>=0.26.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
worker = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
    "nuscenes-devkit>=1.1.11",
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "httpx>=0.26.0",
]

[project.scripts]
mltrain = "src.cli.main:cli"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["B008", "B904", "UP042", "SIM108", "SIM105"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: integration tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
